rename sidebar components
