+++ src/lib/components/RoomSidebar.svelte
